Abstract mock
